pipeline/requirements.txt
anthropic>=0.40.0
boto3>=1.35.0
httpx>=0.27.0
pydantic>=2.0.0
PyGithub>=2.0.0
